pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "svalbard"
version = "0.1.0"
description = "Seed vault for human knowledge — civilization on a stick"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
]

[project.scripts]
svalbard = "svalbard.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/svalbard"]
